requirements.txt
streamlit==1.45.0
faiss-cpu==1.9.0
langchain-openai==0.3.0
langchain-community==0.3.2
openai==1.58.1
pydantic==2.9.2
docx2txt==0.8
pypdf==5.0.1
numpy==1.26.4
python-dotenv==1.0.1